backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1
httpx==0.28.1
python-dotenv==1.1.1
beautifulsoup4==4.13.4
lxml==6.0.0
numpy==2.2.6
psycopg[binary]==3.2.9
python-jose[cryptography]==3.3.0
bcrypt==4.3.0
python-multipart==0.0.20
